Former-commit-id: c2757f91054bf8fe7f97d7b258ce92095adc3afd

+++ backend/project/src/main/java/com/ssafy/project/dto/UserRateDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class UserRateDto {
     private int userNo;
-    private int userRatePoint;
+    private float userRatePoint;
 }
